requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
django-ratelimit==3.0.1
redis==4.5.5
bcrypt==4.0.1
whitenoise==6.5.0